requirements.txt
boto3==1.26.163
botocore==1.29.163
python-dotenv==1.0.0